feat: Implement log file management and viewing features
- Added file handling utilities for opening, clearing, and deleting log files.
- Introduced asynchronous file operations for improved performance.
- Created a log parser to handle log levels and format timestamps.
- Developed a tailing manager for real-time log file monitoring.
- Enhanced user messaging for better feedback on operations.
- Refactored log item representation for better integration with the log viewer.
