+++ src/helpers/messages.ts
@@ -1,0 +1,30 @@
+/**
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ * Copyright (c) 2024-2026 OpenForge Project Contributors
+ */
+
+import * as vscode from "vscode";
+
+// Displays an information message to the user.
+export function showInformationMessage(message: string): void {
+  try {
+    vscode.window.showInformationMessage(message);
+  } catch (error) {
+    console.error(
+      "Failed to show information message:",
+      error instanceof Error ? error.message : String(error),
+    );
+  }
+}
+
+// Displays an error message to the user.
+export function showErrorMessage(message: string): void {
+  try {
+    vscode.window.showErrorMessage(message);
+  } catch (error) {
+    console.error(
+      "Failed to show error message:",
+      error instanceof Error ? error.message : String(error),
+    );
+  }
+}
